requirements.txt
python-dotenv
pyTelegramBotAPI
requests
chromadb
sentence-transformers
duckduckgo-search
PyPDF2
python-docx
